requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-autocomplete-light==3.3.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-reversion==3.0.0
django-tables2==2.0.1
djangorestframework==3.8.2
idna==2.7
isodate==0.6.0
numpy==1.15.2
pandas==0.23.4
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
rdflib==4.2.2
requests==2.19.1
six==1.11.0
SPARQLWrapper==1.8.2
urllib3==1.23
wincertstore==0.2
